rework system-probe flare contents
